refactor: Migrate project to GetX and new architecture
This commit refactors the entire project from BLoC to GetX for state management and adopts a new folder structure.

- Replaced `flutter_bloc` with `get` for state management.
- Replaced `AuthBloc` with `AuthController`.
- Replaced `AuthRepository` and `AuthRemoteDatasource` with a new `AuthService`.
- Renamed `AuthModel` to `UserModel` and moved it to `lib/models`.
- Moved all screens to `lib/screens` and updated them to use the new `AuthController`.
- Updated the dependency injection to use `Get.put()` instead of `get_it`.
- Removed the old `lib/features` directory.
- Added the profile feature as requested in the initial task.
