--- lib/features/auth/domain/models/models.dart
@@ -1,1 +1,0 @@
-export 'auth_model.dart';
