--- lib/features/auth/presentation/auth/auth.dart
@@ -1,5 +1,0 @@
-export 'log_in_screen.dart';
-export 'phone_verification_screen.dart';
-export 'reset_password_screen.dart';
-export 'sign_up_screen.dart';
-export 'widgets/widgets.dart';
