--- lib/features/auth/presentation/auth/widgets/or_divider.dart
@@ -1,26 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-
-class OrDivider extends StatelessWidget {
-  const OrDivider({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 30),
-      child: Row(
-        children: [
-          const Expanded(child: Divider(color: Colors.grey)),
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 15),
-            child: Text(
-              'OR'.tr,
-              style: const TextStyle(fontSize: 18),
-            ),
-          ),
-          const Expanded(child: Divider(color: Colors.grey)),
-        ],
-      ),
-    );
-  }
-}
