--- lib/features/auth/presentation/auth/widgets/widgets.dart
@@ -1,3 +1,0 @@
-export 'media_auth_button.dart';
-export 'media_auth_options.dart';
-export 'or_divider.dart';
