--- lib/features/auth/presentation/bloc/auth_event.dart
@@ -1,68 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-abstract class AuthEvent extends Equatable {
-  const AuthEvent();
-
-  @override
-  List<Object?> get props => [];
-}
-
-class AuthCheckRequested extends AuthEvent {}
-
-class AuthSignUpRequested extends AuthEvent {
-  final String email;
-  final String password;
-  final String name;
-
-  const AuthSignUpRequested(this.email, this.password, this.name);
-
-  @override
-  List<Object?> get props => [email, password];
-}
-
-class AuthLoginRequested extends AuthEvent {
-  final String email;
-  final String password;
-
-  const AuthLoginRequested(this.email, this.password);
-
-  @override
-  List<Object?> get props => [email, password];
-}
-
-class AuthLogoutRequested extends AuthEvent {}
-
-class AuthResetPasswordRequested extends AuthEvent {
-  final String email;
-
-  const AuthResetPasswordRequested(this.email);
-
-  @override
-  List<Object?> get props => [email];
-}
-
-class AuthVerifyPhoneNumberRequested extends AuthEvent {
-  final String phoneNumber;
-
-  const AuthVerifyPhoneNumberRequested(this.phoneNumber);
-
-  @override
-  List<Object?> get props => [phoneNumber];
-}
-
-class AuthVerifyPhoneCodeRequested extends AuthEvent {
-  final String verificationId;
-  final String smsCode;
-
-  const AuthVerifyPhoneCodeRequested(this.verificationId, this.smsCode);
-
-  @override
-  List<Object?> get props => [verificationId, smsCode];
-}
-
-class AuthRest extends AuthEvent {
-  const AuthRest();
-
-  @override
-  List<Object?> get props => [];
-}
