--- lib/features/auth/presentation/intro/widgets/app_description.dart
@@ -1,39 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../../../../shared/widgets/app_icon_widget.dart';
-
-class AppDescription extends StatelessWidget {
-  const AppDescription({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Theme.of(context).scaffoldBackgroundColor.withOpacity(0.8),
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            blurRadius: 10,
-            offset: const Offset(0, 4),
-          ),
-        ],
-      ),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          AppIconWidget(size: 150),
-          const SizedBox(height: 16),
-          Text(
-            'Wellcome to Medical City app',
-            maxLines: 1,
-            style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-          ),
-          const SizedBox(height: 16),
-        ],
-      ),
-    );
-  }
-}
