--- lib/features/auth/presentation/intro/widgets/widgets.dart
@@ -1,3 +1,0 @@
-export 'app_description.dart';
-export 'background_image.dart';
-export 'start_auth_button.dart';
