--- lib/features/main/presentation/widgets/service_card.dart
@@ -1,34 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ServiceCard extends StatelessWidget {
-  final String icon;
-  final String title;
-  const ServiceCard({super.key, required this.icon, required this.title});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        InkWell(
-          onTap: () {},
-          child: Card(
-            elevation: 1,
-            color: Colors.blue[700],
-            child: Container(
-              height: 60,
-              width: 60,
-              padding: EdgeInsets.all(15),
-              child: Image.asset(icon),
-            ),
-          ),
-        ),
-
-        Text(
-          title,
-          textAlign: TextAlign.center,
-          style: const TextStyle(fontSize: 12, fontWeight: FontWeight.w500),
-        ),
-      ],
-    );
-  }
-}
