--- lib/features/redirect/redirect_screen.dart
@@ -1,46 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:go_router/go_router.dart';
-
-import '../../app/router/navigation_routes.dart';
-import '../../shared/widgets/error_screen.dart';
-import '../../shared/widgets/loading_view.dart';
-import '../auth/presentation/bloc/auth_bloc.dart';
-import '../auth/presentation/bloc/auth_event.dart';
-import '../auth/presentation/bloc/auth_state.dart';
-
-class RedirectScreen extends StatelessWidget {
-  const RedirectScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<AuthBloc, AuthState>(
-      builder: (context, state) {
-        if (state is AuthInitial) {
-          WidgetsBinding.instance.addPostFrameCallback((_) {
-            context.read<AuthBloc>().add(AuthCheckRequested());
-          });
-        }
-
-        if (state is AuthLoggedOut) {
-          WidgetsBinding.instance.addPostFrameCallback((_) {
-            // Reset UserBloc when logout is detected
-            // context.read<UserBloc>().add(UserLogout());
-            context.replace(NavigationRoutes.intro);
-          });
-        }
-        if (state is AuthAuthenticated) {
-          WidgetsBinding.instance.addPostFrameCallback((_) {
-            // Ensure UserBloc loads fresh data for the authenticated user
-            // context.read<UserBloc>().add(LoadUser());
-            context.go(NavigationRoutes.main);
-          });
-        }
-        if (state is AuthError) {
-          return ErrorScreen(message: state.message);
-        }
-        return const LoadingView();
-      },
-    );
-  }
-}
